requirements.txt
Flask>=2.3.2
Flask-SQLAlchemy>=2.5.1
Flask-marshmallow>=0.14.0
marshmallow-sqlalchemy>=0.28.1